ozw-frontend: first pass to adapt to new backend
We grabbed this code from our previous backend, which kinda sucked. We
are now implementing a new version, after a few lessons learned (but not
enough, mind you), and we need to adapt the frontend to its new reality.

The problem is that, until we finish, the frontend will be in a sorry
state.

Signed-off-by: Joao Eduardo Luis <joao@wipwd.dev>